gaming/prepa: fixed profile editor commit issue clients_gui: fixed united values spinbox issue

+++ src/libraries/preparation/MineAttributes.h
@@ -72,7 +72,7 @@ private:
     //! @name Member data
     //@{
     kernel::UnitedValue< unsigned int > activityTime_;
-    kernel::UnitedValue< float >        density_;
+    kernel::UnitedValue< double >       density_;
     //@}
 };
